Conflicts:
	llevaMe-eclipse/src/co/edu/eafit/llevame/view/DetallesRutaConductor.java
	llevaMe-eclipse/src/co/edu/eafit/llevame/view/FormularioCrearRuta.java
	llevaMe-eclipse/src/co/edu/eafit/llevame/view/RutasConductor.java
	llevaMe-eclipse/src/co/edu/eafit/llevame/view/RutasPasajero.java

+++ llevaMe-eclipse/src/co/edu/eafit/llevame/model/Ubicacion.java
@@ -3,12 +3,13 @@ package co.edu.eafit.llevame.model;
 public class Ubicacion {
 	private int id;
 	private String nombre;
-	private float latitud;
-	private float longitud;
+	private double latitud;
+	private double longitud;
 	
 	public int getId() {
 		return id;
 	}
+
 	public void setId(int id) {
 		this.id = id;
 	}
@@ -16,19 +17,24 @@ public class Ubicacion {
 	public String getNombre() {
 		return nombre;
 	}
+
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
-	public float getLatitud() {
+
+	public double getLatitud() {
 		return latitud;
 	}
-	public void setLatitud(float latitud) {
+
+	public void setLatitud(double latitud) {
 		this.latitud = latitud;
 	}
-	public float getLongitud() {
+
+	public double getLongitud() {
 		return longitud;
 	}
-	public void setLongitud(float longitud) {
+
+	public void setLongitud(double longitud) {
 		this.longitud = longitud;
 	}
 }
